+++ code/NCheck.Test/CheckerFactory.cs
@@ -1,5 +1,7 @@
 ﻿namespace NCheck.Test
 {
+    using System.Reflection;
+
     using NCheck.Checking;
 
     public class CheckerFactory : NCheck.CheckerFactory
@@ -13,10 +15,12 @@
 
         private void Initialize()
         {
+            // NB Conventions must be before type registrations if they are to apply.
+            Convention(x => typeof(IIdentifiable).IsAssignableFrom(x) ? CompareTarget.Id : CompareTarget.Unknown);
+            Convention((PropertyInfo x) => x.Name == "Ignore" ? CompareTarget.Ignore : CompareTarget.Unknown);
+
             Register(typeof(CheckerFactory).Assembly);
             Register(typeof(NCheck.CheckerFactory).Assembly);
-
-            Builder = new CheckerBuilder(this);
         }
     }
 }