+++ code/NCheck.Test/Checking/SampleClass.cs
@@ -1,0 +1,11 @@
+﻿namespace NCheck.Test.Checking
+{
+    public class SampleClass : IIdentifiable
+    {
+        public object Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Ignore { get; set; }
+    }
+}