+++ code/NCheck.Test/Checking/SampleEnum.cs
@@ -1,0 +1,10 @@
+﻿namespace NCheck.Test.Checking
+{
+    public enum SampleEnum
+    {            
+        KnownKnown = 0,
+        UnknownKnown = 1,
+        KnownUnknown = 2,
+        UnknownUnknown = 3, 
+    }
+}