+++ code/NCheck.Test/Checking/SampleStruct.cs
@@ -1,0 +1,9 @@
+﻿namespace NCheck.Test.Checking
+{
+    public struct SampleStruct
+    {
+        public int A { get; set; }
+
+        public int B { get; set; }
+    }
+}