+++ code/NCheck.Test/Descendant.cs
@@ -1,0 +1,6 @@
+﻿namespace NCheck.Test
+{
+    public class Descendant : Parent
+    {
+    }
+}