+++ code/NCheck/CheckerFactory.cs
@@ -12,7 +12,7 @@
     /// A factory class that provides checking facilities for objects so that property level comparisons can be easily made
     /// </summary>
     public class CheckerFactory : ICheckerFactory
-    {
+    {        
         private readonly IDictionary<Type, IChecker> checkers;
         private ICheckerBuilder checkerBuilder;
 
@@ -24,7 +24,8 @@
             checkers = new Dictionary<Type, IChecker>();
 
             PropertyCheck.IdentityChecker = new NullIdentityChecker();
-            PropertyCheck.Targeter = new TypeCompareTargeter();
+            PropertyCheck.TypeCompareTargeter = new TypeCompareTargeter();
+            PropertyCheck.PropertyCompareTargeter = new PropertyCompareTargeter();
         }
 
         /// <summary>
@@ -32,8 +33,102 @@
         /// </summary>
         public ICheckerBuilder Builder
         {
-            get { return checkerBuilder ?? (checkerBuilder = new NullCheckerBuilder()); }
+            get { return checkerBuilder ?? (checkerBuilder = new CheckerBuilder(this)); }
             set { checkerBuilder = value; }
+        }
+
+        /// <summary>
+        /// Clear the registered checkers
+        /// </summary>
+        public void Clear()
+        {
+            checkers.Clear();
+        }
+
+        /// <copydocfrom cref="ICheckerFactory.Check{T}(IEnumerable{T}, IEnumerable{T}, string)" />
+        public void Check<T>(IEnumerable<T> expectedList, IEnumerable<T> candidateList, string objectName = "")
+        {
+            var checker = Checker<T>();
+            var listChecker = new ListChecker();
+
+            // NB Can't call Check as this retrieves via the property info which we don't have here.
+            listChecker.CheckList(checker as IChecker, expectedList, candidateList, objectName);
+        }
+
+        /// <copydocfrom cref="ICheckerFactory.Check{T}(IEnumerable{T}, IEnumerable{T}, string)" />
+        public void Check<T>(T expected, T candidate, string objectName = "")
+        {
+            if (Verify(candidate, expected, objectName))
+            {
+                Checker<T>().Check(expected, candidate, objectName);
+            }
+        }
+
+        /// <summary>
+        /// General entry point that works out which checker to invoke.
+        /// </summary>
+        /// <param name="expected">Object containing expected values</param>
+        /// <param name="candidate">Object containing values to check</param>
+        /// <param name="objectName">Name of the object, used to disambiguate error messages</param>
+        /// <remarks>
+        /// We use the expected to determine the type of the checker.
+        /// </remarks>        
+        public void Check(object expected, object candidate, string objectName = "")
+        {
+            if (expected == null && candidate == null)
+            {
+                return;
+            }
+
+            var type = expected != null ? expected.GetType() : candidate.GetType();
+            Checker(type).Check(expected, candidate, objectName);
+        }
+
+        /// <copydocfrom cref="ICheckerFactory.CheckParent{T}" />
+        public void CheckParent<T>(T expected, T candidate, string objectName = "")
+        {
+            if (Verify(candidate, expected, objectName))
+            {
+                Checker<T>().CheckBase(expected, candidate, objectName);
+            }
+        }
+
+        /// <summary>
+        /// Register a <see cref="CompareTarget"/> convention for a type.
+        /// </summary>
+        /// <typeparam name="T">Type to use</typeparam>
+        /// <param name="target">CompareTarget value to return</param>
+        public void Convention<T>(CompareTarget target)
+        {
+            Convention(typeof(T), target);
+        }
+
+        /// <summary>
+        /// Register a <see cref="CompareTarget"/> convention for a type.
+        /// </summary>
+        /// <param name="type">Type to use</param>
+        /// <param name="target">CompareTarget value to return</param>
+        public void Convention(Type type, CompareTarget target)
+        {
+            PropertyCheck.TypeCompareTargeter.Register(type, target);
+        }
+
+        /// <summary>
+        /// Register a <see cref="CompareTarget"/> convention based on type information.
+        /// </summary>
+        /// <param name="convention"></param>
+        public void Convention(Func<Type, CompareTarget> convention)
+        {
+            PropertyCheck.TypeCompareTargeter.Register(convention);
+        }
+
+        /// <summary>
+        /// Register a <see cref="CompareTarget"/> convention based on property information.
+        /// </summary>
+        /// <param name="convention"></param>
+        public void Convention(Func<PropertyInfo, CompareTarget> convention)
+        {
+            PropertyCheck.PropertyCompareTargeter.Register(convention);
         }
 
         /// <summary>
@@ -89,80 +184,6 @@
             Add(typeof(T), (IChecker)checker);
 
             return this;
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expectedList"></param>
-        /// <param name="candidateList"></param>
-        /// <param name="objectName"></param>
-        public void Check<T>(IEnumerable<T> expectedList, IEnumerable<T> candidateList, string objectName)
-        {
-            var checker = Checker<T>();
-            var listChecker = new ListChecker();
-
-            // NB Can't call Check as this retrieves via the property info which we don't have here.
-            listChecker.CheckList(checker as IChecker, expectedList, candidateList, objectName);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        public void Check<T>(T expected, T candidate)
-        {
-            Check(expected, candidate, string.Empty);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        /// <param name="objectName"></param>
-        public void Check<T>(T expected, T candidate, string objectName)
-        {
-            if (Verify(candidate, expected, objectName))
-            {
-                Checker<T>().Check(expected, candidate, objectName);
-            }
-        }
-
-        /// <summary>
-        /// General entry point that works out which checker to invoke
-        /// </summary>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        /// <param name="objectName"></param>
-        public void Check(object expected, object candidate, string objectName)
-        {
-            if (expected == null && candidate == null)
-            {
-                return;
-            }
-
-            var type = expected != null ? expected.GetType() : candidate.GetType();
-            Checker(type).Check(expected, candidate, objectName);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        /// <param name="objectName"></param>
-        public void CheckParent<T>(T expected, T candidate, string objectName)
-        {
-            if (Verify(candidate, expected, objectName))
-            {
-                Checker<T>().CheckBase(expected, candidate, objectName);
-            }
         }
 
         /// <summary>
@@ -236,12 +257,13 @@
                 return checker;
             }
 
-            throw new ArgumentOutOfRangeException(string.Format("No checker registered for {0}", type.FullName));
+            throw new NotSupportedException(string.Format("No checker registered for {0}", type.FullName));
         }
 
         private void Add(Type entityType, IChecker checker)
         {
-            checkers.Add(entityType, checker);
+            // Use assignment so we can override an existing checker.
+            checkers[entityType] = checker;
         }
 
         private class ListChecker : PropertyCheck