+++ code/NCheck/ICheckerFactory.cs
@@ -11,19 +11,11 @@
         /// <summary>
         /// Check that the properties of two instances of <see typeparamref="T" /> are equal
         /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        void Check<T>(T expected, T candidate);
-
-        /// <summary>
-        /// Check that the properties of two instances of <see typeparamref="T" /> are equal
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        /// <param name="objectName"></param>
-        void Check<T>(T expected, T candidate, string objectName);
+        /// <typeparam name="T">Type to check</typeparam>
+        /// <param name="expected">Expected object to use</param>
+        /// <param name="candidate">Candidate object to use</param>
+        /// <param name="objectName">Name to use, displayed in error messages to disambiguate</param>
+        void Check<T>(T expected, T candidate, string objectName = "");
 
         /// <summary>
         /// Check if two collections of <see typeparamref="T" /> are equal.
@@ -34,22 +26,22 @@
         /// also <see cref="ICollection{T}" /> which means that it supports <see cref="ICollection{T}.Count" />
         /// If these checks are passed, each item is compared in turn.
         /// </remarks>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expectedList"></param>
-        /// <param name="candidateList"></param>
-        /// <param name="objectName"></param>
-        void Check<T>(IEnumerable<T> expectedList, IEnumerable<T> candidateList, string objectName);
+        /// <typeparam name="T">Type to check</typeparam>
+        /// <param name="expectedList">Expected object to use</param>
+        /// <param name="candidateList">Candidate object to use</param>
+        /// <param name="objectName">Name to use, displayed in error messages to disambiguate</param>
+        void Check<T>(IEnumerable<T> expectedList, IEnumerable<T> candidateList, string objectName = "");
 
         /// <summary>
         /// Check that the properties of the parent class of <see typeparamref="T" /> are equal.
         /// </summary>
         /// <remarks>
-        /// Only used when testing a class which is part of a hiearchy
+        /// Only used when testing a class which is part of a hierarchy
         /// </remarks>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="expected"></param>
-        /// <param name="candidate"></param>
-        /// <param name="objectName"></param>
-        void CheckParent<T>(T expected, T candidate, string objectName);
+        /// <typeparam name="T">Type to check</typeparam>
+        /// <param name="expected">Expected object to use</param>
+        /// <param name="candidate">Candidate object to use</param>
+        /// <param name="objectName">Name to use, displayed in error messages to disambiguate</param>
+        void CheckParent<T>(T expected, T candidate, string objectName = "");
     }
 }