+++ RELEASE_NOTES.md
@@ -1,2 +1,8 @@
+#### 1.1 (2014-06-15)
+* 1.1: Change default builder to CheckerBuilder from NullCheckerBuilder
+Refactor TypeCompareTargeter to be more modular and allow registration of custom functions.
+Introduced IPropertyCompareTargeter/PropertyCompareTargeter to allow for property-based overrides
+Introduced CheckerFactory.Convention methods as helpers for registering both type and property based overrides
+
 #### 1.0 (2014-04-16)
 * 1.0: First release.